+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/pausemenu/states/StateMananger.java
@@ -3,12 +3,16 @@ package de.mephisto.vpin.commons.fx.pausemenu.states;
 import de.mephisto.vpin.commons.fx.pausemenu.MenuController;
 import de.mephisto.vpin.commons.fx.pausemenu.PauseMenu;
 import de.mephisto.vpin.commons.fx.pausemenu.UIDefaults;
+import de.mephisto.vpin.restclient.games.GameRepresentation;
+import de.mephisto.vpin.restclient.games.GameStatus;
 import de.mephisto.vpin.restclient.popper.PinUPControl;
 import de.mephisto.vpin.restclient.popper.PinUPControls;
+import de.mephisto.vpin.restclient.popper.PopperScreen;
 import javafx.application.Platform;
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
 import javafx.util.Duration;
+import org.jnativehook.GlobalScreen;
 import org.jnativehook.keyboard.NativeKeyEvent;
 import org.jnativehook.keyboard.NativeKeyListener;
 import org.slf4j.Logger;
@@ -25,12 +29,13 @@ public class StateMananger implements NativeKeyListener {
   private MenuState activeState;
 
   private final static StateMananger INSTANCE = new StateMananger();
-  private boolean blocked;
 
   private int LEFT;
   private int RIGHT;
   private int ENTER;
   private int BACK;
+
+  private MenuController menuController;
 
   public static StateMananger getInstance() {
     return INSTANCE;
@@ -59,46 +64,32 @@ public class StateMananger implements NativeKeyListener {
     });
   }
 
-  public void setInputBlocked(boolean b, long duration) {
-    this.blocked = b;
-    new Thread(() -> {
-      try {
-        Thread.sleep(duration);
-      } catch (InterruptedException e) {
-        //ignore1
-      }
-      this.blocked = !b;
-    }).start();
-  }
-
-  public void setInputBlocked(boolean unblock) {
-    this.blocked = unblock;
-  }
-
   public void init(MenuController controller) {
+    this.menuController = controller;
     this.activeState = new MenuItemSelectionState(controller);
   }
 
   public void handle(int keyCode) {
-    if (blocked) {
-      return;
-    }
-
     if (keyCode == LEFT) {
-      navPlayer.play();
+      if (menuController.isAtStart()) {
+        return;
+      }
+
       this.activeState = activeState.left();
-      setInputBlocked(true, UIDefaults.SELECTION_SCALE_DURATION);
+      navPlayer.play();
     }
     else if (keyCode == RIGHT) {
-      navPlayer.play();
+      if (menuController.isAtEnd()) {
+        return;
+      }
+
       this.activeState = activeState.right();
-      setInputBlocked(true, UIDefaults.SELECTION_SCALE_DURATION);
+      navPlayer.play();
     }
     else if (keyCode == ENTER) {
       enterPlayer.play();
       this.activeState = activeState.enter();
       LOG.info("Entered " + this.activeState);
-      setInputBlocked(true, UIDefaults.SELECTION_SCALE_DURATION);
     }
     else if (keyCode == BACK) {
       PauseMenu.exit();
@@ -111,14 +102,19 @@ public class StateMananger implements NativeKeyListener {
 
   @Override
   public void nativeKeyPressed(NativeKeyEvent nativeKeyEvent) {
+    GlobalScreen.removeNativeKeyListener(StateMananger.getInstance());
     Platform.runLater(() -> {
       handle(nativeKeyEvent.getRawCode());
+      new Thread(() -> {
+        try {
+          Thread.sleep(UIDefaults.SELECTION_SCALE_DURATION);
+          GlobalScreen.addNativeKeyListener(StateMananger.getInstance());
+        } catch (InterruptedException e) {
+          throw new RuntimeException(e);
+        }
+      }).start();
     });
-    try {
-      Thread.sleep(UIDefaults.SELECTION_SCALE_DURATION);
-    } catch (InterruptedException e) {
-      throw new RuntimeException(e);
-    }
+
   }
 
   @Override
@@ -130,5 +126,14 @@ public class StateMananger implements NativeKeyListener {
     RIGHT = pinUPControls.getKeyCode(PinUPControl.FUNCTION_GAME_NEXT);
     ENTER = pinUPControls.getKeyCode(PinUPControl.FUNCTION_GAME_START);
     BACK = pinUPControls.getKeyCode(PinUPControl.FUNCTION_EXIT);
+  }
+
+  public void setGame(GameRepresentation game, GameStatus status, PopperScreen screen) {
+    GlobalScreen.addNativeKeyListener(StateMananger.getInstance());
+    menuController.setGame(game, status, screen);
+  }
+
+  public void exit() {
+    GlobalScreen.removeNativeKeyListener(StateMananger.getInstance());
   }
 }