+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ As a result, your existing **"UI Settings" and "Server Settings"** in the prefer
 - Studio Window: The size and location of the Studio window is now persisted and restored on the next launch.
 - File Dialogs: The last used folder location is now stored and used as initial directory for every new file selection dialog.
 - Added "Reload" button to MAME section, to invalidate the MAME settings cache.
+- Improved server start-up time a bit.
 - Re-worked the preferences for the highscore card pop-up:
   - Renamed this feature from "notification" to "pop-up".
   - Added the option to use a PinUP player screen configuration for the pop-up positioning, instead of just displaying it on top of the playfield.